bot/requirements.txt
fastapi
uvicorn
discord.py
mcrcon
psycopg2-binary
pandas
prometheus-client
python-dotenv
